back/pyproject.toml
[tool.poetry]
name = "hestia"
version = "0.1.0"
description = ""
authors = ["David Francos <opensource@davidfrancos.net>"]

[tool.poetry.dependencies]
python = "^3.8"
aiohttp-swagger3 = "^0.5.0"
cleo = "^0.8.1"
gql = {version = "^3.0.0-alpha.3", allow-prereleases = true}

[tool.poetry.dev-dependencies]
yapf = "^0.30.0"
watchgod = "^0.6"

[tool.poetry.scripts]
hestia = 'hestia:main'
[build-system]
requires = ["poetry>=0.12"]
build-backend = "poetry.masonry.api"

